Refactored layout choice
