+++ packages/frontend/web/pages/Article.tsx
@@ -10,7 +10,7 @@ import { Section } from '@frontend/web/components/Section';
 import { Nav } from '@frontend/web/components/Nav/Nav';
 import { HeaderAdSlot } from '@root/packages/frontend/web/components/HeaderAdSlot';
 
-import { DecideLayout } from './DecideLayout';
+import { DecideLayout } from '../layouts/DecideLayout';
 import { CAPI } from '@root/fixtures/CAPI';
 
 export const Article: React.FC<{
